- контроллер для реализации фукнционала логина через соц. сеть сделан универсальным - удаление данных из таблиц бд реализовано с помощью ajax запросов ( страница не обновляется при удалении) - методы создания и редактирования контроллеров разделены, а также созданы валидированные запросы - фикс багов

+++ app/Http/Controllers/admin/UserController.php
@@ -32,15 +32,9 @@ class UserController extends Controller
         ]);
     }
 
-    public function delete(User $user)
-    {
-        return view('admin.users.delete', ['user' => $user]);
-    }
-
     public function store(Request $request)
     {
-        if($request->isMethod('post'))
-        {
+        if ($request->isMethod('post')) {
             $request->merge([
                 'is_admin' => $request->has('is_admin'),
             ]);
@@ -52,71 +46,72 @@ class UserController extends Controller
                 'is_admin' => 'nullable|boolean',
             ]);
 
-            // хеширование пароля перед сохранением
             $validatedData['password'] = bcrypt($validatedData['password']);
-
 
             try {
                 $user = User::create($validatedData);
-            }catch (\Exception $e){
-                return redirect()->route('admin.users.create')->with('error', 'Ошибка добавления пользователя '
-                    . $e->getMessage());
+            } catch (\Exception $e) {
+                return redirect()->route('admin.users.create')->with('error', 'Ошибка добавления пользователя ' . $e->getMessage());
             }
 
-            return redirect()->route('admin.users.index', $user->id)->with('success','Пользователь успешно добавлен');
+            return redirect()->route('admin.users.index')->with('success', 'Пользователь успешно добавлен');
         }
+    }
 
-        if($request->isMethod('put'))
-        {
-            $user = User::findOrFail($request->id);
-
-            // проверка, чтобы не изменять имя или поле is_admin у пользователя с именем "Admin" или у самого себя
+    public function update(UpdateUserRequest $request, User $user)
+    {
+        if ($request->isMethod('put')) {
             if ($user->name == 'Admin' || $user->id == auth()->id()) {
-                // если имя Admin или это текущий пользователь, сбрасываем изменения
                 $request->merge([
-                    'name' => $user->name, // оставляем имя неизменным
-                    'is_admin' => $user->is_admin, // оставляем поле is_admin неизменным
+                    'name' => $user->name,
+                    'is_admin' => $user->is_admin,
                 ]);
                 $message = $user->name == 'Admin' ? 'Невозможно изменить имя и роль Администратор для пользователя с именем Admin.' : 'Невозможно изменить роль Администратор для самого себя.';
                 session()->flash('warning', $message);
-            }
-            $validatedData = $request->validate([
-                'name' => 'required|string|max:255',
-                'email' => 'required|string|email|max:255|unique:users,email,' . $request->id,
-                'password' => 'nullable|string|min:8|confirmed',
-                'is_admin' => 'nullable|boolean',
-            ]);
 
-            // проверка и хеширование пароля
+                return redirect()->route('admin.users.edit', $user->id)->with('error', $message);
+            }
+
+            $validatedData = $request->validated();
+
+            $validatedData['is_admin'] = $request->has('is_admin') ? true : false;
+
             if ($request->filled('password')) {
                 $validatedData['password'] = bcrypt($validatedData['password']);
             } else {
-                // если пароль не был передан, удаляем его из массива данных
                 unset($validatedData['password']);
             }
 
-
             try {
-                $user = User::query()->find($request->id);
                 $user->update($validatedData);
-            }catch (\Exception $e){
-                return redirect()->route('admin.users.edit')->with('error', 'Ошибка обновления пользователя '
-                    . $e->getMessage());
+            } catch (\Exception $e) {
+                return redirect()->route('admin.users.edit', $user->id)->with('error', 'Ошибка обновления пользователя ' . $e->getMessage());
             }
-            return redirect()->route('admin.users.index',$user->id)->with('success','Пользователь успешно обновлен');
+
+            return redirect()->route('admin.users.index')->with('success', 'Пользователь успешно обновлен');
         }
     }
 
-    public function deleteUser(Request $request)
+    public function delete(User $user)
     {
+        if ($user->name == 'Admin' || $user->id == auth()->id()) {
+            return response()->json([
+                'error' => 'Невозможно удалить пользователя с именем Admin или самого себя.'
+            ], 400); // Ошибка с кодом 400
+        }
+
         try {
-            $user = User::findOrFail($request->id);
             $user->delete();
         } catch (\Exception $e) {
-            return redirect()->route('admin.users.index')->with('error', 'Ошибка удаления пользователя ' . $e->getMessage());
+            return response()->json([
+                'error' => 'Ошибка удаления пользователя: ' . $e->getMessage()
+            ], 500);
         }
 
-        return redirect()->route('admin.users.index')->with('success', 'Пользователь успешно удален');
+        return response()->json([
+            'success' => 'Пользователь успешно удален'
+        ]);
     }
+
 }
